Fixed two more Visual C++ compiler warnings
git-svn-id: file:///srv/devel/repo-conversion/nusu@457 d2e56fa2-650e-0410-a79f-9358c0239efd

+++ Benchmarks/Events/LSignalBenchmark.cpp
@@ -152,7 +152,7 @@ namespace Nuclex { namespace Support { namespace Events {
 
   BENCHMARK_F(Invoke2_x100, LSignal, Event2Fixture, 1000, 0) {
     for(std::size_t index = 0; index < 100; ++index) {
-      this->testEvent(index);
+      this->testEvent(static_cast<int>(index));
     }
   }
 
@@ -160,7 +160,7 @@ namespace Nuclex { namespace Support { namespace Events {
 
   BENCHMARK_F(Invoke50_x100, LSignal, Event50Fixture, 1000, 0) {
     for(std::size_t index = 0; index < 100; ++index) {
-      this->testEvent(index);
+      this->testEvent(static_cast<int>(index));
     }
   }
